PolyFox/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
gunicorn
python-dotenv
requests
boto3